source/server/requirements.txt
django==1.7.1
django-cors-headers==1.0.0
djangorestframework==3.0.2
pillow==2.6.1
django-oauth-toolkit==0.7.2